# contracts/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "crusher-api",
    "version": "0.3.0"
  },
  "paths": {
...
    "/api/v1/analyses": {
      "post": {
        "summary": "Create Analysis",
        "description": "提交分析任务；真实模式下拒绝 demo_error。",
        "operationId": "create_analysis_api_v1_analyses_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateAnalysisRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CreateAnalysisResponse"
                }
              }
            }
          },
          "400": {
            "description": "业务拒绝或输入超长",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "请求参数不合法",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/analyses/{task_id}": {
      "get": {
        "summary": "Get Analysis",
        "description": "按 task_id 查询状态与原文；失败任务不返回报告。",
        "operationId": "get_analysis_api_v1_analyses__task_id__get",
        "parameters": [
          {
            "name": "task_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Task Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TaskResponse"
                }
              }
            }
          },
          "404": {
            "description": "任务不存在",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "请求参数不合法",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/dual-analyses": {
      "post": {
        "summary": "Create Dual Analysis",
        "description": "销售话术 vs 正式材料对照（同步返回，不走任务轮询）。",
        "operationId": "create_dual_analysis_api_v1_dual_analyses_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DualAnalysisRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DualAnalysisReport"
                }
              }
            }
          },
          "400": {
            "description": "材料不足或超长",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "请求参数不合法",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/documents/extract": {
      "post": {
        "summary": "Extract Document",
        "description": "上传 PDF/图片并提取可编辑文本；用户确认后再进入金融分析。",
        "operationId": "extract_document_api_v1_documents_extract_post",
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_extract_document_api_v1_documents_extract_post"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ExtractedDocument"
                }
              }
            }
          },
          "400": {
            "description": "文件不合法或超限",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "请求参数不合法",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            }
          },
          "503": {
            "description": "OCR 不可用",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "AnalysisReport": {
        "properties": {
          "product_candidates": {
            "items": {
              "$ref": "#/components/schemas/ProductCandidate"
            },
            "type": "array",
            "title": "Product Candidates"
          },
          "resolved_product_type": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ProductTypeId"
              },
              {
                "type": "null"
              }
            ]
          },
          "analysis_scope": {
            "$ref": "#/components/schemas/AnalysisScope",
            "default": "supported"
          },
          "scope_reason": {
            "type": "string",
            "title": "Scope Reason",
            "default": ""
          },
          "product_risk_grade": {
            "$ref": "#/components/schemas/ProductRiskGrade"
          },
          "plain_language": {
            "$ref": "#/components/schemas/PlainLanguage"
          },
          "key_parameters": {
            "items": {
              "$ref": "#/components/schemas/KeyParameter"
            },
            "type": "array",
            "title": "Key Parameters"
          },
          "findings": {
            "items": {
              "$ref": "#/components/schemas/Finding"
            },
            "type": "array",
            "title": "Findings"
          },
          "missing_disclosures": {
            "items": {
              "$ref": "#/components/schemas/MissingDisclosure"
            },
            "type": "array",
            "title": "Missing Disclosures"
          },
          "general_references": {
            "items": {
              "$ref": "#/components/schemas/GeneralReference"
            },
            "type": "array",
            "title": "General References"
          },
          "pending_questions": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Pending Questions",
            "description": "待确认问题（可与 missing_disclosures 互补）"
          },
          "disclaimer": {
            "type": "string",
            "minLength": 1,
            "title": "Disclaimer",
            "default": "本 Demo 不进行用户适当性评估，不构成投资建议。"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "product_risk_grade",
          "plain_language"
        ],
        "title": "AnalysisReport",
        "description": "完整分析报告。"
      },
      "AnalysisScope": {
        "type": "string",
        "enum": [
          "supported",
          "out_of_scope",
          "needs_confirmation"
        ],
        "title": "AnalysisScope",
        "description": "一次分析的范围状态。"
      },
      "ApiErrorDetail": {
        "properties": {
          "error_code": {
            "type": "string",
            "minLength": 1,
            "title": "Error Code"
          },
          "message": {
            "type": "string",
            "minLength": 1,
            "title": "Message"
          },
          "fields": {
            "anyOf": [
              {
                "items": {
                  "type": "string"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Fields"
          },
          "max_input_chars": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Max Input Chars"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "error_code",
          "message"
        ],
        "title": "ApiErrorDetail",
        "description": "稳定 4xx 错误体（与 HTTP 实际 payload 一致）。"
      },
      "ApiErrorResponse": {
        "properties": {
          "detail": {
            "$ref": "#/components/schemas/ApiErrorDetail"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "detail"
        ],
        "title": "ApiErrorResponse",
        "description": "FastAPI HTTPException / 校验失败统一外层。"
      },
      "Body_extract_document_api_v1_documents_extract_post": {
        "properties": {
          "files": {
            "items": {
              "type": "string",
              "contentMediaType": "application/octet-stream"
            },
            "type": "array",
            "title": "Files",
            "description": "PDF 或图片，可多文件"
          }
        },
        "type": "object",
        "required": [
          "files"
        ],
        "title": "Body_extract_document_api_v1_documents_extract_post"
      },
...
      "ClaimComparison": {
        "properties": {
          "comparison_id": {
            "type": "string",
            "minLength": 1,
            "title": "Comparison Id"
          },
          "subject": {
            "$ref": "#/components/schemas/ClaimSubject"
          },
          "status": {
            "$ref": "#/components/schemas/ClaimStatus"
          },
          "summary": {
            "type": "string",
            "minLength": 1,
            "title": "Summary"
          },
          "sales_claim": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/Claim"
              },
              {
                "type": "null"
              }
            ]
          },
          "official_evidence": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/EvidenceRef"
              },
              {
                "type": "null"
              }
            ]
          },
          "suggested_follow_up": {
            "type": "string",
            "title": "Suggested Follow Up",
            "default": ""
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "comparison_id",
          "subject",
          "status",
          "summary"
        ],
        "title": "ClaimComparison",
        "description": "一张对照卡：销售主张 vs 正式材料。"
      },
...
      "CreateAnalysisRequest": {
        "properties": {
          "text": {
            "type": "string",
            "maxLength": 8000,
            "minLength": 1,
            "title": "Text",
            "description": "待分析原文"
          },
          "product_hint": {
            "$ref": "#/components/schemas/ProductHint",
            "default": "auto"
          },
          "locale": {
            "type": "string",
            "const": "zh-CN",
            "title": "Locale",
            "default": "zh-CN"
          },
          "demo_error": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/DemoErrorKind"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "text"
        ],
        "title": "CreateAnalysisRequest",
        "description": "POST /api/v1/analyses 唯一请求体（应用层与 HTTP 共用）。\n\nJava 对照：创建分析的 Request DTO；字段与 OpenAPI 同源。\n业务不变量：text 去空白后非空；product_hint 仅允许首版枚举；locale 仅 zh-CN。"
      },
      "CreateAnalysisResponse": {
        "properties": {
          "task_id": {
            "type": "string",
            "title": "Task Id"
          },
          "task_status": {
            "$ref": "#/components/schemas/TaskStatus"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "task_id",
          "task_status"
        ],
        "title": "CreateAnalysisResponse",
        "description": "创建分析任务的响应。"
      },
      "DemoErrorKind": {
        "type": "string",
        "enum": [
          "model_timeout",
          "invalid_json",
          "rate_limited"
        ],
        "title": "DemoErrorKind"
      },
      "DualAnalysisReport": {
        "properties": {
          "sales_source": {
            "$ref": "#/components/schemas/SourceDocument"
          },
          "official_source": {
            "$ref": "#/components/schemas/SourceDocument"
          },
          "comparisons": {
            "items": {
              "$ref": "#/components/schemas/ClaimComparison"
            },
            "type": "array",
            "title": "Comparisons"
          },
          "pending_questions": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Pending Questions"
          },
          "disclaimer": {
            "type": "string",
            "title": "Disclaimer",
            "default": "本对照仅比较已提交材料中的表述，不判断合同法律效力，不对销售人员打分。"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "sales_source",
          "official_source"
        ],
        "title": "DualAnalysisReport",
        "description": "双材料对照报告。"
      },
      "DualAnalysisRequest": {
        "properties": {
          "sales_text": {
            "type": "string",
            "maxLength": 8000,
            "minLength": 1,
            "title": "Sales Text"
          },
          "official_text": {
            "type": "string",
            "maxLength": 8000,
            "minLength": 1,
            "title": "Official Text"
          },
          "product_hint": {
            "$ref": "#/components/schemas/ProductHint",
            "default": "auto"
          },
          "locale": {
            "type": "string",
            "const": "zh-CN",
            "title": "Locale",
            "default": "zh-CN"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "sales_text",
          "official_text"
        ],
        "title": "DualAnalysisRequest",
        "description": "POST /api/v1/dual-analyses 请求体。"
      },
      "ErrorCode": {
        "type": "string",
        "enum": [
          "INPUT_TOO_LONG",
          "MODEL_TIMEOUT",
          "RATE_LIMITED",
          "INVALID_MODEL_JSON",
          "RULE_FAILED",
          "KNOWLEDGE_UNAVAILABLE",
          "TASK_NOT_FOUND",
          "FORBIDDEN_CLIENT_CONFIG",
          "OCR_UNAVAILABLE",
          "DOCUMENT_PARSE_FAILED",
          "INTERNAL_ERROR"
        ],
        "title": "ErrorCode"
      },
...
      "ExtractedDocument": {
        "properties": {
          "document_id": {
            "type": "string",
            "title": "Document Id"
          },
          "filename": {
            "type": "string",
            "minLength": 1,
            "title": "Filename"
          },
          "media_type": {
            "type": "string",
            "minLength": 1,
            "title": "Media Type"
          },
          "overall_status": {
            "$ref": "#/components/schemas/StageStatus"
          },
          "pages": {
            "items": {
              "$ref": "#/components/schemas/PageExtractResult"
            },
            "type": "array",
            "title": "Pages"
          },
          "combined_text": {
            "type": "string",
            "title": "Combined Text",
            "default": ""
          },
          "sensitive_hints": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Sensitive Hints",
            "description": "低置信或敏感数字提示，供 H5 人工确认"
          },
          "message": {
            "type": "string",
            "title": "Message",
            "default": ""
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "filename",
          "media_type",
          "overall_status"
        ],
        "title": "ExtractedDocument",
        "description": "用户确认前的提取结果（可编辑文本）。"
      },
...
      "PageExtractResult": {
        "properties": {
          "page": {
            "type": "integer",
            "minimum": 1.0,
            "title": "Page"
          },
          "status": {
            "$ref": "#/components/schemas/PageExtractStatus"
          },
          "text": {
            "type": "string",
            "title": "Text",
            "default": ""
          },
          "confidence": {
            "type": "number",
            "maximum": 1.0,
            "minimum": 0.0,
            "title": "Confidence",
            "default": 1.0
          },
          "error_message": {
            "type": "string",
            "title": "Error Message",
            "default": ""
          },
          "used_ocr": {
            "type": "boolean",
            "title": "Used Ocr",
            "default": false
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "page",
          "status"
        ],
        "title": "PageExtractResult"
      },
...
      "PlainLanguage": {
        "properties": {
          "text": {
            "type": "string",
            "minLength": 1,
            "title": "Text"
          },
          "status": {
            "$ref": "#/components/schemas/StageStatus",
            "default": "success"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "text"
        ],
        "title": "PlainLanguage"
      },
      "ProductCandidate": {
        "properties": {
          "product_type_id": {
            "$ref": "#/components/schemas/ProductTypeId"
          },
          "product_type_name": {
            "type": "string",
            "minLength": 1,
            "title": "Product Type Name"
          },
          "confidence": {
            "type": "number",
            "maximum": 1.0,
            "minimum": 0.0,
            "title": "Confidence"
          },
          "evidence_quotes": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Evidence Quotes"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "product_type_id",
          "product_type_name",
          "confidence"
        ],
        "title": "ProductCandidate"
      },
      "ProductHint": {
        "type": "string",
        "enum": [
          "auto",
          "structured_deposit",
          "loan"
        ],
        "title": "ProductHint",
        "description": "接口允许的手动产品提示（首版仅自动 / 结构性存款 / 贷款）。"
      },
...
      "ProductTypeId": {
        "type": "string",
        "enum": [
          "structured_deposit",
          "loan",
          "snowball",
          "insurance",
          "fund",
          "unknown"
        ],
        "title": "ProductTypeId"
      },
      "SourceDocument": {
        "properties": {
          "source_id": {
            "type": "string",
            "title": "Source Id"
          },
          "source_type": {
            "$ref": "#/components/schemas/SourceType"
          },
          "name": {
            "type": "string",
            "minLength": 1,
            "title": "Name"
          },
          "text": {
            "type": "string",
            "maxLength": 8000,
            "minLength": 1,
            "title": "Text"
          },
          "pages": {
            "type": "integer",
            "minimum": 1.0,
            "title": "Pages",
            "default": 1
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "source_type",
          "name",
          "text"
        ],
        "title": "SourceDocument",
        "description": "一次对照中的一侧材料。"
      },
      "SourceType": {
        "type": "string",
        "enum": [
          "sales_pitch",
          "official_document",
          "user_input"
        ],
        "title": "SourceType"
      },
      "StageInfo": {
        "properties": {
          "name": {
            "type": "string",
            "title": "Name"
          },
          "status": {
            "$ref": "#/components/schemas/StageStatus"
          },
          "message": {
            "type": "string",
            "title": "Message",
            "default": ""
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "name",
          "status"
        ],
        "title": "StageInfo"
      },
      "StageStatus": {
        "type": "string",
        "enum": [
          "success",
          "partial",
          "failed",
          "not_applicable"
        ],
        "title": "StageStatus",
        "description": "某一个分析步骤的状态。"
      },
      "TaskResponse": {
        "properties": {
          "task_id": {
            "type": "string",
            "title": "Task Id"
          },
          "task_status": {
            "$ref": "#/components/schemas/TaskStatus"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time",
            "title": "Updated At"
          },
          "stages": {
            "items": {
              "$ref": "#/components/schemas/StageInfo"
            },
            "type": "array",
            "title": "Stages"
          },
          "error_code": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ErrorCode"
              },
              {
                "type": "null"
              }
            ]
          },
          "error_message": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error Message"
          },
          "report": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/AnalysisReport"
              },
              {
                "type": "null"
              }
            ]
          },
          "is_failure": {
            "type": "boolean",
            "title": "Is Failure",
            "default": false
          },
          "source_text": {
            "type": "string",
            "title": "Source Text",
            "default": ""
          },
          "product_hint": {
            "$ref": "#/components/schemas/ProductHint",
            "default": "auto"
          },
          "resolved_product_type": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ProductTypeId"
              },
              {
                "type": "null"
              }
            ]
          },
          "analysis_scope": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/AnalysisScope"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "task_id",
          "task_status",
          "created_at",
          "updated_at",
          "stages"
        ],
        "title": "TaskResponse",
        "description": "查询任务状态/报告的响应（含任务绑定原文）。"
      },
      "TaskStatus": {
        "type": "string",
        "enum": [
          "queued",
          "running",
          "completed",
          "failed"
        ],
        "title": "TaskStatus",
        "description": "整次分析任务：排队 / 进行中 / 成功 / 失败。"
      }
    }
  }
}